FFmpeg の実行部分を別クラスに分離
